Change information types. years are ints and lastPlayed, firstPlayed to QDateTime

+++ src/playlist/view/listview/PrettyItemDelegate.cpp
@@ -793,7 +793,7 @@ void Playlist::PrettyItemDelegate::setModelData( QWidget * editor, QAbstractItem
                     break;
                 }
             case Year:
-                ec->setYear( value );
+                ec->setYear( value.toInt() );
                 break;
             case Bpm:
                 ec->setBpm( value.toFloat() );
